Dockerfile
This file contains something called a Docker file. This is a text file that includes a set of instructions that is used to build a Docker image. A Docker image puts everything into a container that is used to run the application usually via the cloud. It is important to have clear and concise steps in the Dockerfile to ensure consistency when building a web application.